% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lin.R
\name{lin}
\alias{lin}
\title{Graph Linear Regression and store Data}
\usage{
lin(varx, vary)
}
\arguments{
\item{varx}{A vector of numbers.}

\item{vary}{A vector of numbers.}
}
\value{
Plot the two vectors on a scatter plot along with the linear regression model, then store and return the int, slope, R, and R^2 values.
}
\description{
Graph Linear Regression and store Data
}
\examples{
lin(c(1,2,3), c(2,4,6))
}

