% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat201-package.r
\docType{package}
\name{stat201-package}
\alias{stat201-package}
\title{Stat201}
\description{
TODO
}
\details{
\tabular{ll}{
   Package: \tab textgateway \cr
   Type: \tab Package \cr
   License: \tab BSD 2-clause \cr
}
}
\author{
Drew Schmidt
}
\references{
Project home page: \url{https://github.com/wrathematics/stat201}
}
\keyword{Package}

